global-price-sentinel/requirements.txt
playwright==1.40.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.4
jinja2==3.1.2
pydantic==2.5.2
pydantic-settings==2.1.0
apscheduler==3.10.4
requests==2.31.0
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
email-validator==2.1.0.post1
